move EdisonHalAppender to scm-core because we need it for plugins
